clang and size_t

+++ tasks/mpi/titov_s_simple_iteration/include/ops_mpi.hpp
@@ -50,7 +50,7 @@ class MPISimpleIterationParallel : public ppc::core::Task {
   std::vector<double> Values;
   std::vector<double> current;
   std::vector<double> prev;
-  size_t Rows;
+  int Rows;
 
   double epsilon_;
